chore: fix all merge conflicts compile and run program
