+++ server/relay/src/constants.rs
@@ -2,6 +2,3 @@
 
 /// Queue for the manager to receive all tasks
 pub static RELAY_QUEUE: &str = "relay_queue";
-
-// Exchange for the manager to publish tasks to
-pub static RELAY_EXCHANGE: &str = "relay_exchange";
