+++ server/relay/src/testing.rs
@@ -1,11 +1,9 @@
 #[cfg(test)]
 pub mod test {
-    use std::sync::Arc;
-
     use axum_test::TestServer;
     use sqlx::PgPool;
 
-    use crate::{brokers::core::MockBrokerProducer, lifecycle::setup_app};
+    use crate::lifecycle::setup_app;
 
     pub static MIGRATOR: sqlx::migrate::Migrator = sqlx::migrate!();
 
@@ -21,7 +19,7 @@ pub mod test {
     /// Creates and returns a test server instance with the application router.
     /// This provides a way to make test HTTP requests against the API endpoints.
     pub async fn get_test_server(db_pools: PgPool) -> TestServer {
-        let app = setup_app(&db_pools, Arc::new(MockBrokerProducer::new())).await;
+        let app = setup_app(&db_pools).await;
         TestServer::new(app).unwrap()
     }
 }
